+++ tests/network-tests/run-test-scenario.sh
@@ -9,5 +9,6 @@ SCENARIO=$1
 # fallback if scenario not specified
 SCENARIO=${SCENARIO:="full"}
 
-# Execute the tests
-time DEBUG=integration-tests:* yarn workspace network-tests node-ts-strict src/scenarios/${SCENARIO}.ts
+# Execute the scenario and optional arguments after following scenario name
+time DEBUG=integration-tests:* yarn workspace network-tests \
+    node-ts-strict src/scenarios/${SCENARIO}.ts $2 $3 $4 $5