+++ tests/network-tests/run-tests.sh
@@ -4,6 +4,8 @@ set -e
 SCRIPT_PATH="$(dirname "${BASH_SOURCE[0]}")"
 cd $SCRIPT_PATH
 
+rm ./output.json || :
+
 function cleanup() {
     docker logs joystream-node --tail 15 || :
     docker stop joystream-node || :
@@ -25,29 +27,16 @@ sleep 30
 # Display runtime version
 yarn workspace api-scripts tsnode-strict src/status.ts | grep Runtime
 
-# Start any other services we want
-# docker-compose -f ../../docker-compose.yml up -d colossus-1
-
 # Start a query-node
 if [ "${NO_QN}" != true ]
 then
   ../../query-node/start.sh
 fi
 
-# Execute tests
-
 if [ "${NO_STORAGE}" != true ]
 then
   ./start-storage.sh
-  export REUSE_KEYS=true
-  export SKIP_STORAGE_AND_DISTRIBUTION=true
 fi
 
-# First scenario..
-IGNORE_HIRED_LEADS=true ./run-test-scenario.sh $1
-
-# In between pickup generated keys from first scenario or bootstrap scene with all well known
-# keys for workers and members..
-
-# Second scenario..
-# ./run-test-scenario.sh $2
+# Execute tests
+./run-test-scenario.sh $*