+++ tests/network-tests/src/fixtures/membership/UpdateProfileHappyCaseFixture.ts
@@ -16,7 +16,6 @@ export type MemberProfileData = {
   about?: string | null
   avatarUri?: string | null
   externalResources?: MembershipMetadata.IExternalResource[] | null
-  validatorAccount?: string | null
 }
 
 export class UpdateProfileHappyCaseFixture extends BaseQueryNodeFixture {
@@ -57,7 +56,6 @@ export class UpdateProfileHappyCaseFixture extends BaseQueryNodeFixture {
     )
     assert.isFalse(Utils.hasDuplicates(metadata.externalResources?.map(({ type }) => type)))
     assert.equal(metadata.isVerifiedValidator, false)
-    assert.equal(metadata.validatorAccount, expected.validatorAccount)
   }
 
   public getExpectedValues(): MemberProfileData {
@@ -69,9 +67,6 @@ export class UpdateProfileHappyCaseFixture extends BaseQueryNodeFixture {
       externalResources: isSet(this.newValues.externalResources)
         ? this.newValues.externalResources || null
         : this.oldValues.externalResources,
-      validatorAccount: isSet(this.newValues.validatorAccount)
-        ? this.newValues.validatorAccount || null
-        : this.oldValues.validatorAccount,
     }
   }
 
@@ -107,7 +102,6 @@ export class UpdateProfileHappyCaseFixture extends BaseQueryNodeFixture {
       about: this.newValues.about,
       avatarUri: this.newValues.avatarUri,
       externalResources: this.newValues.externalResources,
-      validatorAccount: this.newValues.validatorAccount,
     })
     this.tx = this.api.tx.members.updateProfile(
       this.memberContext.memberId,